+++ .pre-commit-config.yaml
@@ -1,0 +1,37 @@
+repos:
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v2.3.0
+    hooks:
+      - id: check-yaml
+        args: [--allow-multiple-documents]
+      - id: check-json
+      - id: trailing-whitespace
+  - repo: https://github.com/psf/black
+    rev: 24.8.0
+    hooks:
+      - id: black
+        exclude: "venv|vendor|__jac_gen__|generated"
+  - repo: https://github.com/PyCQA/flake8
+    rev: 6.1.0
+    hooks:
+      - id: flake8
+        args: ["--config=jac/.flake8"]
+        additional_dependencies:
+          [
+            pep8-naming,
+            flake8_import_order,
+            flake8_comprehensions,
+            flake8_bugbear,
+            flake8_annotations,
+            flake8_simplify,
+          ]
+        exclude: "venv|vendor|__jac_gen__|generated|examples|langserve/tests|pygame_mock"
+  - repo: https://github.com/pre-commit/mirrors-mypy
+    rev: v1.15.0
+    hooks:
+      - id: mypy
+        exclude: "venv|vendor|__jac_gen__|generated|examples|tests|stubs|support|setup.py|jac-mtllm"
+        args:
+          - --follow-imports=silent
+          - --ignore-missing-imports
+        additional_dependencies: [types-PyYAML]
