+++ jac-cloud/jac_cloud/core/memory.py
@@ -6,7 +6,7 @@ from typing import Callable, Generator, Iterable, TypeVar, cast
 
 from bson import ObjectId
 
-from jaclang.plugin.feature import JacFeature as Jac
+from jaclang.runtimelib.feature import JacFeature as Jac
 from jaclang.runtimelib.memory import Memory
 
 
@@ -93,7 +93,7 @@ class MongoDB(Memory[ObjectId, BaseAnchor | Anchor]):
         """Find one anchor from memory by ids with filter."""
         return next(self.find(anchors, filter, session), None)
 
-    def find_by_id(self, anchor: BA) -> BA | None:
+    def find_by_id(self, anchor: BA) -> BaseAnchor | Anchor | None:
         """Find one by id."""
         data = super().find_by_id(anchor.id)
