+++ jac-cloud/jac_cloud/plugin/cli.py
@@ -7,8 +7,8 @@ from pickle import load
 from typing import Any
 
 from jaclang.cli.cmdreg import cmd_registry
-from jaclang.plugin.default import hookimpl
 from jaclang.runtimelib.context import ExecutionContext
+from jaclang.runtimelib.default import hookimpl
 from jaclang.runtimelib.machine import JacMachine, JacProgram
 
 from pymongo.errors import ConnectionFailure, OperationFailure
