+++ jac-cloud/jac_cloud/plugin/implementation/api.py
@@ -28,7 +28,7 @@ from fastapi import (
 )
 from fastapi.responses import ORJSONResponse
 
-from jaclang.plugin.feature import JacFeature as Jac
+from jaclang.runtimelib.feature import JacFeature as Jac
 
 from orjson import loads
