+++ jac/jaclang/__init__.py
@@ -1,7 +1,7 @@
 """The Jac Programming Language."""
 
-from jaclang.plugin.default import JacFeatureImpl
-from jaclang.plugin.feature import JacFeature, plugin_manager
+from jaclang.runtimelib.default import JacFeatureImpl
+from jaclang.runtimelib.feature import JacFeature, plugin_manager
 
 
 plugin_manager.register(JacFeatureImpl)
