+++ jac/jaclang/cli/cli.py
@@ -17,11 +17,11 @@ from jaclang.compiler.constant import Constants
 from jaclang.compiler.passes.main.pyast_load_pass import PyastBuildPass
 from jaclang.compiler.passes.main.schedules import py_code_gen_build, py_code_gen_typed
 from jaclang.compiler.program import JacProgram
-from jaclang.plugin.builtin import dotgen
-from jaclang.plugin.feature import JacCmd as Cmd
-from jaclang.plugin.feature import JacFeature as Jac
+from jaclang.runtimelib.builtin import dotgen
 from jaclang.runtimelib.constructs import WalkerArchitype
 from jaclang.runtimelib.context import ExecutionContext
+from jaclang.runtimelib.feature import JacCmd as Cmd
+from jaclang.runtimelib.feature import JacFeature as Jac
 from jaclang.runtimelib.machine import JacMachine
 from jaclang.utils.helpers import debugger as db
 from jaclang.utils.lang_tools import AstTool
