+++ jac/jaclang/runtimelib/tests/fixtures/graph_purger.jac
@@ -1,4 +1,4 @@
-import:py from jaclang.plugin.feature {JacFeature as Jac}
+import:py from jaclang.runtimelib.feature {JacFeature as Jac}
 
 node A {
     has id: int;
