+++ scripts/check.sh
@@ -1,0 +1,2 @@
+pre-commit run --files $(find jac -type f)
+pre-commit run --files $(find jac-cloud -type f)
