+++ tool_plugins/streamlit/jaclang_streamlit/commands.py
@@ -5,7 +5,7 @@ import sys
 import tempfile
 
 from jaclang.cli.cmdreg import cmd_registry
-from jaclang.plugin.default import hookimpl
+from jaclang.runtimelib.default import hookimpl
 
 import streamlit.web.bootstrap as bootstrap
